Update chellenge_7.cpp
add float to challenge 7

+++ week_01/MyWork/Challenge 7/chellenge_7.cpp
@@ -3,11 +3,11 @@ int main(){
 	int a, b;
 	scanf("%d%d", &a, &b);
 	
-	printf("a+b : %d \n", a+b);
-	printf("a-b : %d\n", a-b);
-	printf("a*b : %d\n", a*b);
+	printf("a+b : %d \n", (float) a+b);
+	printf("a-b : %d\n", (float) a-b);
+	printf("a*b : %d\n", (float) a*b);
 	printf("a/b : %f\n", (float) a/b);
-	printf("a%b : %d\n", a%b);
+	printf("a%b : %d\n", (float) a%b);
 	
 	
 	return 0;
